extract page edit and add
