+++ test/e2e/utils/ui/index.js
@@ -2,5 +2,6 @@ export { clickHeaderBtnMySurveys, clickHeaderBtnCreateSurvey } from './header'
 export { clickHomeBtnEditSurveyInfo, expectHomeDashboard } from './home'
 export { clickSidebarBtnHome, clickSidebarBtnSurveyForm } from './sidebar'
 export { waitForLoader } from './loader'
-export { expectItemIsTheLastNodeDef, expectItemsAreInOrderAsNodeDef } from './surveyForm'
+export { expectItemIsTheLastNodeDef, expectItemsAreInOrderAsNodeDef, expectSurveyFormLoaded } from './surveyForm'
 export { addItemToPage } from './nodeDefDetail'
+export { editPage, addSubPage } from './surveyFormPage'
